refactor: split global notification center
